requirements.txt
websocket-client
martypy
